requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.4.0
# cx_Oracle==8.3.0          # Uncomment for Oracle
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1
python-multipart==0.0.9
